+++ src/test/java/com/marginallyclever/ro3/TestAllPanels.java
@@ -1,22 +1,16 @@
 package com.marginallyclever.ro3;
 
 import com.formdev.flatlaf.FlatLightLaf;
-import com.marginallyclever.ro3.node.Node;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledIfEnvironmentVariable;
-import org.reflections.Reflections;
-import org.reflections.util.ClasspathHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.swing.*;
-import java.util.List;
 
 /**
  * <p>Display a card layout of all the {@link JPanel}s in the project.  This would be handy for translators to see all
  * the panels in one place.</p>
- * <p>{@link Reflections} can only find classes that extend {@link JPanel}.  Therefore in a second pass create one
- * instance of every {@link Node} and add the components from {@link Node#getComponents(List)}.</p>
  */
 @DisabledIfEnvironmentVariable(named = "CI", matches = "true")
 public class TestAllPanels {
@@ -27,16 +21,9 @@ public class TestAllPanels {
         Registry.start();
 
         // print the classpath
-        System.out.println("Classpath:");
-        for(var path : ClasspathHelper.forJavaClassPath()) {
-            System.out.println("  "+path);
-        }
+        System.out.println("Classpath: "+ System.getProperty("java.class.path"));
 
-        try {
-            UIManager.setLookAndFeel(new FlatLightLaf());
-        } catch (Exception ignored) {
-            logger.error("Failed to set look and feel.");
-        }
+        FlatLightLaf.setup();
 
         AllPanels allPanels = new AllPanels();
         var frame = allPanels.buildFrame();